cloud_computing/requirements.txt
google-cloud-pubsub==2.20.0

